+++ test/integration/notify.test.ts
@@ -1,9 +1,11 @@
 import {
   AddressActivityWebhook,
   Alchemy,
+  CustomGraphqlWebhook,
   Network,
   NftActivityWebhook,
   NftFilter,
+  NftMetadataUpdateWebhook,
   WebhookType
 } from '../../src';
 import { loadAlchemyEnv } from '../test-util';
@@ -29,9 +31,12 @@ describe('E2E integration tests', () => {
   ];
 
   const webhookUrl = 'https://temp-site.ngrok.io';
+  const graphqlQuery = '{ block { hash } }';
 
   let addressWh: AddressActivityWebhook;
   let nftWh: NftActivityWebhook;
+  let nftMetadataWh: NftMetadataUpdateWebhook;
+  let customWh: CustomGraphqlWebhook;
 
   async function createInitialWebhooks(): Promise<void> {
     addressWh = await alchemy.notify.createWebhook(
@@ -43,6 +48,16 @@ describe('E2E integration tests', () => {
       webhookUrl,
       WebhookType.NFT_ACTIVITY,
       { filters: nftFilters, network: Network.ETH_MAINNET }
+    );
+    nftMetadataWh = await alchemy.notify.createWebhook(
+      webhookUrl,
+      WebhookType.NFT_METADATA_UPDATE,
+      { filters: nftFilters, network: Network.ETH_MAINNET }
+    );
+    customWh = await alchemy.notify.createWebhook(
+      webhookUrl,
+      WebhookType.GRAPHQL,
+      { graphqlQuery, network: Network.ETH_MAINNET }
     );
   }
 
@@ -62,6 +77,13 @@ describe('E2E integration tests', () => {
     const all = await alchemy.notify.getAllWebhooks();
     expect(all.totalCount).toBeGreaterThan(0);
     expect(all.webhooks.length).toEqual(all.totalCount);
+  });
+
+  it('getGraphqlQuery()', async () => {
+    let response = await alchemy.notify.getGraphqlQuery(customWh);
+    expect(response.graphqlQuery).toEqual(graphqlQuery);
+    response = await alchemy.notify.getGraphqlQuery(customWh.id);
+    expect(response.graphqlQuery).toEqual(graphqlQuery);
   });
 
   it('getAddresses()', async () => {
@@ -193,6 +215,27 @@ describe('E2E integration tests', () => {
     ).toEqual(0);
   });
 
+  it('create and delete CustomWebhook', async () => {
+    const customWebhook = await alchemy.notify.createWebhook(
+      webhookUrl,
+      WebhookType.GRAPHQL,
+      { graphqlQuery, network: Network.ETH_GOERLI }
+    );
+    expect(customWebhook.url).toEqual(webhookUrl);
+    expect(customWebhook.type).toEqual(WebhookType.GRAPHQL);
+    expect(customWebhook.network).toEqual(Network.ETH_GOERLI);
+    let response = await alchemy.notify.getAllWebhooks();
+    expect(
+      response.webhooks.filter(wh => wh.id === customWebhook.id).length
+    ).toEqual(1);
+
+    await alchemy.notify.deleteWebhook(customWebhook);
+    response = await alchemy.notify.getAllWebhooks();
+    expect(
+      response.webhooks.filter(wh => wh.id === customWebhook.id).length
+    ).toEqual(0);
+  });
+
   it('create AddressActivityWebhook with ENS', async () => {
     const rawAddress = '0xd8dA6BF26964aF9D7eEd9e03E53415D37aA96045';
     const addressWebhook = await alchemy.notify.createWebhook(
@@ -226,12 +269,35 @@ describe('E2E integration tests', () => {
     ).toEqual(0);
   });
 
+  it('create and delete NftActivityWebhook', async () => {
+    const nftActivityWebhook = await alchemy.notify.createWebhook(
+      webhookUrl,
+      WebhookType.NFT_METADATA_UPDATE,
+      { filters: nftFilters, network: Network.ETH_GOERLI }
+    );
+    expect(nftActivityWebhook.url).toEqual(webhookUrl);
+    expect(nftActivityWebhook.type).toEqual(WebhookType.NFT_METADATA_UPDATE);
+    expect(nftActivityWebhook.network).toEqual(Network.ETH_GOERLI);
+    let response = await alchemy.notify.getAllWebhooks();
+    expect(
+      response.webhooks.filter(wh => wh.id === nftActivityWebhook.id).length
+    ).toEqual(1);
+
+    await alchemy.notify.deleteWebhook(nftActivityWebhook.id);
+    response = await alchemy.notify.getAllWebhooks();
+    expect(
+      response.webhooks.filter(wh => wh.id === nftActivityWebhook.id).length
+    ).toEqual(0);
+  });
+
   it('update NftActivityWebhook filter with same filter', async () => {
     const addFilters = [
+      // Duplicate filter
       {
         contractAddress: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
         tokenId: '234'
       },
+      // New Filter
       {
         contractAddress: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
         tokenId: '123'
@@ -270,6 +336,71 @@ describe('E2E integration tests', () => {
     const updated = response.webhooks.filter(wh => wh.id === nftWh.id);
     expect(updated.length).toEqual(1);
     expect(updated[0].isActive).toEqual(false);
+  });
+
+  it('update CustomWebhook status', async () => {
+    const webhooks = await alchemy.notify.getAllWebhooks();
+    const filteredWebhooks = webhooks.webhooks.filter(
+      webhook => webhook.id === customWh.id
+    );
+    expect(filteredWebhooks.length).toEqual(1);
+    const customWebhook = filteredWebhooks[0];
+    const currStatus = customWebhook.isActive;
+    await alchemy.notify.updateWebhook(customWebhook.id, {
+      isActive: !currStatus
+    });
+    const response = await alchemy.notify.getAllWebhooks();
+    const updated = response.webhooks.filter(wh => wh.id === customWebhook.id);
+    expect(updated.length).toEqual(1);
+    expect(updated[0].isActive).toEqual(!currStatus);
+  });
+
+  it('update NftMetadataUpdateWebhook status', async () => {
+    await alchemy.notify.updateWebhook(nftMetadataWh.id, {
+      isActive: false
+    });
+    const response = await alchemy.notify.getAllWebhooks();
+    const updated = response.webhooks.filter(wh => wh.id === nftMetadataWh.id);
+    expect(updated.length).toEqual(1);
+    expect(updated[0].isActive).toEqual(false);
+  });
+
+  it('update NftMetadataUpdateWebhook filter with same filter', async () => {
+    const addMetadataFilters = [
+      // Duplicate filter
+      {
+        contractAddress: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+        tokenId: '234'
+      },
+      // New Filter
+      {
+        contractAddress: '0x88b48f654c30e99bc2e4a1559b4dcf1ad93fa656',
+        tokenId: '123'
+      }
+    ];
+
+    const removeMetadataFilters = [
+      {
+        contractAddress: '0x17dc95f9052f86ed576af55b018360f853e19ac2',
+        tokenId: 345
+      }
+    ];
+
+    await alchemy.notify.updateWebhook(nftWh, {
+      addFilters: addMetadataFilters,
+      removeFilters: removeMetadataFilters
+    });
+
+    const response = await alchemy.notify.getNftFilters(nftWh);
+    expect(response.filters.length).toEqual(2);
+
+    await alchemy.notify.updateWebhook(nftWh, {
+      removeFilters: removeMetadataFilters
+    });
+
+    await alchemy.notify.updateWebhook(nftWh, {
+      addFilters: addMetadataFilters
+    });
   });
 
   it('update AddressActivityWebhook address', async () => {