+++ test/test-util.ts
@@ -234,7 +234,7 @@ export function createRawNftSale(
 export function createRawContractForOwner(
   address: string,
   tokenId: string,
-  media: Media,
+  media: Media[],
   isSpam?: boolean,
   name?: string,
   tokenType?: NftTokenType,
@@ -242,7 +242,8 @@ export function createRawContractForOwner(
   totalSupply?: string,
   opensea?: RawOpenSeaCollectionMetadata,
   contractDeployer?: string,
-  deployedBlockNumber?: number
+  deployedBlockNumber?: number,
+  title = 'NFT Title'
 ): RawContractForOwner {
   return {
     address,
@@ -252,6 +253,7 @@ export function createRawContractForOwner(
     totalBalance: 1,
     numDistinctTokensOwned: 1,
     name,
+    title,
     totalSupply,
     opensea,
     symbol,
@@ -265,14 +267,16 @@ export function createNftMediaData(
   bytes?: number,
   format?: string,
   thumbnail?: string
-): Media {
-  return {
-    raw: 'http://api.nikeape.xyz/ipfs/nickbanc/1.jpg',
-    gateway: 'http://api.nikeape.xyz/ipfs/nickbanc/1.jpg',
-    bytes,
-    format,
-    thumbnail
-  };
+): Media[] {
+  return [
+    {
+      raw: 'http://api.nikeape.xyz/ipfs/nickbanc/1.jpg',
+      gateway: 'http://api.nikeape.xyz/ipfs/nickbanc/1.jpg',
+      bytes,
+      format,
+      thumbnail
+    }
+  ];
 }
 
 export function verifyNftContractMetadata(